artikels CRUD in artikelsbeheren.jsp

+++ src/main/resources/schema.sql
@@ -27,8 +27,8 @@ CREATE TABLE ARTIKEL
 (
     id          INT         NOT NULL AUTO_INCREMENT PRIMARY KEY,
     naam        VARCHAR(50) NOT NULL,
-    prijs       FLOAT       NOT NULL,
+    prijs       DECIMAL     NOT NULL,
     beschikbaar BIT         NOT NULL,
-/*    opmerking   VARCHAR(100) NULLABLE,*/
+--     opmerking   VARCHAR(100),
     PRIMARY KEY (id)
 );
